Prefer shell variables for environment input in GHA.
If a value is already set in the env, we can access it natively in shell
instead of using `${{ env.KEY }}`. The latter is done after parsing the
yaml and before it is passed to shell script and requires careful
escaping.

These should all be straightforward fixes where we either were in
javascript or shell and were allowing values to be interpolated in via
`${{ env.KEY }}`.